Moved alert dialog to DateSelectDialogFragment
